pyproject.toml
[project]
name = "novel-similarity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "gensim>=4.3.3",
    "jieba>=0.42.1",
    "jupyter>=1.1.1",
    "numpy>=1.18.5",
    "openai>=1.72.0",
    "pip>=25.0.1",
    "pydantic~=1.10.2",
    "pyldavis>=3.4.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
]
